Move BuildsTable
