+++ site/src/components/Workspace/BuildsTable/BuildAvatar.tsx
