+++ site/src/components/Workspace/Workspace.tsx
@@ -11,7 +11,7 @@ import { useTranslation } from "react-i18next"
 import { useNavigate } from "react-router-dom"
 import * as TypesGen from "../../api/typesGenerated"
 import { Alert, AlertDetail } from "../Alert/Alert"
-import { BuildsTable } from "../BuildsTable/BuildsTable"
+import { BuildsTable } from "./BuildsTable/BuildsTable"
 import { Margins } from "../Margins/Margins"
 import { Resources } from "../Resources/Resources"
 import { Stack } from "../Stack/Stack"
